Make it easy to query for regression tests
Summary:
With this change and the addition of the `--replace-all` flag, we an now run all integration tests with a single command.

It will take long, but it is better than waiting to see failed signals in the diff.

We can create a small script that first asks buck to build them all, which it will do in parallel, and then we run all of them in parallel.

Reviewed By: nhawkes

Differential Revision: D48117510

fbshipit-source-id: 2b7d87441bf5e7181f171c3d6116b458ce42e138