Revert D48184154: Multisect successfully blamed "D48184154: [glass] Update some symbol ids that tickle decl/defn symbols" for test or build failures
Summary:
This diff is reverting D48184154
D48184154: [glass] Update some symbol ids that tickle decl/defn symbols by donsbot has been identified to be causing the following test or build failures:

Tests affected:
- [glean/glass/test/regression:cpp - glass-regression-cpp : platform010 - relatedNeighborhood_test_cpp__h_.decl](https://www.internalfb.com/intern/test/281475073005881/)

Here's the Multisect link:
https://www.internalfb.com/multisect/2716682
Here are the tasks that are relevant to this breakage:

We're generating a revert to back out the changes in this diff, please note the backout may land if someone accepts it.

If you believe this diff has been generated in error you may Commandeer and Abandon it.

Reviewed By: donsbot

Differential Revision: D48201650

fbshipit-source-id: 680084aa726a7a537c7755feabb011d462d775c6