Index inherited members
Summary:
This diff augments the class, interface, and trait definitions to contain references to their inherited members (through clusters based on where the members are coming from).

The goals are:
- This is the source of truth and prevents discrepancies between the ad hoc decl folding algorithm Glass uses and what the typechecker thinks. See T152981029 for a user report where the Glass' interpretation was wrong.
- CodeHub pages OOM because on demand decl folding is expensive.
- CodeHub pages take a long time to load because decl folding is expensive.
- Open up use cases for substituting type parameters with arguments using the information in the decl store (currently inherited decls might contain undefined type parameters).

This affects incremental indexing, so now the hash of a file includes the origins of the members of classes in the file.

This diff also adds a next configuration to `hh` called `symbol_write_index_inherited_members` which control whether to enable inherited member indexing. This is set to true by default, but added here just in case we need to roll the feature back. The gating works by not collecting any inherited members in `Symbol_file_info` this in turn turns the folds used for container declarations and hash computation into noops (as we fold over the empty list).

Reviewed By: phlalx

Differential Revision: D46829790

fbshipit-source-id: 71c211fc04a6b938d25c74885729e16c676414fc