Use the already-computed SHA1 for buck-out file digest.
Summary:
This actually simplifies the `glean/facebook/lang/clang:path` module. It doesn't need to read the file again, and doesn't compute the MD5 hash anymore.

We already read the memory buffer to compute the SHA1 hash for the file digest now, so just feed that into the buck-out path mangling.

Reviewed By: donsbot

Differential Revision: D48276283

fbshipit-source-id: 8f6cb9ef277fa2a38a3ddde93997db52faa5beee