Remove 200M calls / week to src.FileLines
Summary:
Since we _always_ check for this we can batch it up with:
- src.File
- codemarkup.IndexedFile
- digest.FileDigest

to build a single FileInfo result in one call. We just have to memoize the line ending to range conversion correctly.

To be clear: these calls are cheap. p98 is 873 compile + execution time.
But we make 200M of them every week.

https://fburl.com/scuba/glean_server/1lj5bg7a

p99 is 10ms. About 500 qps is just these calls.

https://fburl.com/scuba/glean_server/gdzjl88e

Requires glass.FileInfo and friends on the server before landing.

Reviewed By: phlalx

Differential Revision: D48307228

fbshipit-source-id: 4ce6d6ec5b918d0b42525e3b4e9721d76d582d17