Fix typechecking of nested if statement
Summary:
If a nested if statement had a subquery without a result in the condition, we would get a type error stating that the last statement in the subquery should be an expression.

This behaviour was incorrect since we don't expect return values from if conditions.
This diff fixes this check.

Reviewed By: malanka

Differential Revision: D48390474

fbshipit-source-id: 022cd746e5a11a58465901dd9663a7d7ea1fd22d