Include glass symbolId in USRSymbolDefinition responses
Summary:
The problem: enabling the relations method for grpc. This can be done with searchRelated but we need a way to quickly get a glass Sym for a clang USR.
The Solution: Always return the glass sym in a USRSymbolDefinition response And just have the proxy call that first before calling searchRelated method.

Reviewed By: donsbot

Differential Revision: D48413194

fbshipit-source-id: dc8eee1927a6222872e95f7fc72bb8bc8fbff54f