Extend documentSymbol() calls with a table of filepath -> content digest for xref targets
Summary:
Just plumbing and specifying the type for xref content digests.

We will give the clients an option to fetch the source blob from either the buck2 CAS cache.
Currently we only assume the content is in source control, addressable with (repo , filepath) pairs.
However, we have the content digetsts now in at least one language (C++), so we can also generate references to file content not in source control (e.g. temporary build artifacts in buck0-out).

Client can decide whether to use the local filepath or the digest key to the CAS.

Alternatives would be:
- unique integer for every filepath and key by that nat (adds additional processing overhead)
- map<string, FileDigest>, doesn't work with repo targets in other repos (e.g. www/hack in fbsource) , so we have to have the repo prefix encoding.

Reviewed By: tdupes

Differential Revision: D48343967

fbshipit-source-id: 709093d242709899f44856111e87c7b4cc73fcb9