[glean parallelize derivations automatically
Summary:
Makes `glean derive` able of figuring out the right ordering and parallelism of a set of predicate derivations. It builds a graph of predicate derivations and then maps it into a graph of interlocked threads, one for each derivation. Since Haskell has lightweight threads, this has negligible cost and is as simple as it gets.

A nice property is that all the changes are client-side, so this works for both local and remote backends

Reviewed By: lazamar

Differential Revision: D48739470

fbshipit-source-id: 03967970678ce4432f07f7bb5e8cd5f2bf6f813c