More precise exception for strict exact revision requests
Summary:
The hope is that, by marking this exception as stateful and ~~server~~ client, we are giving the ThriftCache enough information to *not* cache it. Or if that is not enough, we can add it to the RP "appErrors" to turn off Thrift caching without it getting marked as a fatal error. See this [comment](https://fb.workplace.com/groups/1730279463893632/posts/2980366948884871/?comment_id=2981647258756840&reply_comment_id=2986411611613738).

If this works, we may want to mark `GlassException` itself as "stateful client" to prevent ThriftCache from caching it.

Reviewed By: donsbot

Differential Revision: D48863372

fbshipit-source-id: e4bef8cc4d5ddf7325d3b415fab1976f773b4316