Declare RevisionNotAvailable exception in service methods
Summary:
D48863372 added this exception type which can be thrown currently from `documentSymbolListX` and `documentSymbolIndex`, but forgot to add it to the list of declared exceptions for those service methods. This is causing Thrift fatals, caught by Conveyour staged deployment checks.

Adding it to the declared exceptions of all the service methods that take a `RequestOptions` input since all those are susceptible of throwing a `RevisionNotAvailable` exception.

Reviewed By: donsbot

Differential Revision: D48944440

fbshipit-source-id: e200478de2434dfe1b90abdd839912742c59c1fc