Add field for explicit derived fact dependencies
Summary:
To support externally derived predicates, let's allow clients to specify which facts were required to derive some given fact.
We are adding this to `Batch`, which is used by `ComputedBatch`, so it means that as a start we will only support this functionality with binary batches.

This encoding is flexible enough to allow us to support recursive predicates down the line.

Reviewed By: pepeiborra

Differential Revision: D48518385

fbshipit-source-id: 016fc9fbf9a8b4b1f161bcfefdf8622653b306a0