Track ownership of multiple predicate in DefineOwnership
Summary:
Currently `DefineOwnership` is used when deriving a single predicate and thus
restricts ownership information to facts of that predicate.

We want to support the case where we track onwership for facts of multiple
predicates in the same batch.

This would be useful for external derivers and for the derivation of
co-recursive predicates.

Reviewed By: pepeiborra

Differential Revision: D48914195

fbshipit-source-id: f1ce755c0f19f8d2ff4ec6207b94d40b6a17f403