Create DefineOwnership from ComputedBatch
Summary:
When the client sends us information about which facts depend on which, we need to create a `DefineOwnership` value to hold this information.
But when should we do that?

A `DefineOwnership` doesn't hold a relationship between facts and facts, but a relationship between facts and owners.
This means that if B depends on A, and A has owner O1, B will also have owner O1.
But if later A also becomes owned by O2, B will not inherit this owner.

Therefore we should wait until we know we will find all possible owners of the dependencies to only then translate the fact-fact mapping into the fact-owner mapping.
We do that at the time of writing the batch.

This is accomplished by doing two things:
- Ensure that externally derived facts can only depend on facts from complete predicates
- Create the `DefineOwnership` after all possible deduplication has been done.

Point one here presents a problem for co-recursive predicates, but we don't have those for now.

Reviewed By: pepeiborra

Differential Revision: D48914196

fbshipit-source-id: 6b9c6f3bc334076c951f11eb4333e74ab53a06f3