Emit xrefs for indirect uses `via` if they are in the local file
Summary:
We have xrefs that are uses pointing to indirect targets. The targets can be `using` declarations. This is essentially the xref for the using declaration site. We already have all the data in glass:
- pull out the indirect "via" and 'target" and use that to build an xref

Basicaly, we're flattening the chains of indirect xrefs for using declarations to emit one xref for each link in the chain.

To start with: just one level of xrefs. Since via chains can go into other files, we have to be careful to anchor the source.Range to the relevant file.

Reviewed By: pepeiborra

Differential Revision: D48961769

fbshipit-source-id: 24f218aee8cd0238e6bc3735b01641056c9e0c85