Return entities even if their destination (name span) isn't present
Summary:
We happened to not have the destination name spans for usingDirectives and usingDeclarations.

The effect of that is these decl kinds get filtered out of the file entity list. Work around this with a `if` check on the presence of the (optional) name span. This will have no cost and makes the code slightly more robust to degrading name spans.

I suspect this is part of UPND due to not having an xref target or hover for things like

https://pxl.cl/3jS4F

I note the symbol ids of these usingDirectives could probably do with work as they're just fqnames without a tag.

Reviewed By: pepeiborra

Differential Revision: D48959361

fbshipit-source-id: 18728139118646ad8f19be5d1fdb98502cc57629