Accept multiple files with -out FILE
Summary:
It's weird that multiple files are only accepted when `-output FILE` is not given. As a result, snapshots are computed differently for prod and testing, with different perf characteristics.

Here we remove this restriction, change the meaning to `-output FOLDER` and write every snapshot to the sensible place, and propagate the change to our Skycastle script

This is safe to deploy because:

- `-output` is not used in prod, only for testing
- The change is behaviour preserving

Reviewed By: phlalx

Differential Revision: D49011552

fbshipit-source-id: 544a87a46f7651b6ae829761c995ce165de48c09