Only fail with `IndexFailure` on `hasUncompilableErrorOccurred`.
Summary:
There are a bunch of `IndexFailure` facts in the `fbsource` DB where 0 errors are captured. This is because we use the `hasErrorOccurred` as the condition for bailing, but we actually only care about error and fatals. `hasErrorOccurred` is true also for warnings propagated to error, and things like that. This diff uses `hasUncompilableErrorOccurred` instead.

```
fbsource> :count src.IndexFailure _

17159 results, 17159 facts, 29.64ms, 14761584 bytes, 584 compiled bytes

fbsource> :count src.IndexFailure { details = "0 error(s) occurred while indexing" }

13115 results, 13115 facts, 21.12ms, 6530968 bytes, 780 compiled bytes
```

Querying over the [Indexer Coverage](https://fburl.com/scuba/indexer_coverage/m5h3ylgh) table shows that fixing this issue will index ~3,500 more files directly in `fbsource`, and potentially more indirectly affected files. But the actual user impact will likely be lower, most of the 3,500 files are already indexed in other DBs. There are still [173 files](P829689600) that are not indexed in other DBs.

Reviewed By: malanka

Differential Revision: D49318003

fbshipit-source-id: f23e7ec351e553f8771a64c24e47e41505bae1df