query servers should recover from fetch backups errors
Summary:
The janitor fetches the list of remote DBs periodically. In the event of an error, we fall back to the empty list. This implies that `glean list` will only include local DB results, and can lead to missing DB SEVs like S365963 or to stale DBs getting used (the local ones).

There are two alternatives worth considering:
1. Reuse the last good list of remote DBs
2. Die

This Diff implements 1, but there is some more work to do to prevent the server from eternally reusing the last good list of remote DBs. Perhaps failing a HC until there is a successful fetch, or dying after 5 failed attempts.

Reviewed By: malanka

Differential Revision: D49534465

fbshipit-source-id: 093de86ff2854a4d2730ff8f24849c946b49cd36