filter available DBs before computing retention set
Summary:
We can compute the retention set with full precision if we know which DBs are available elsewhere. That simplifies code a bit as we dont' need to track the remote retention set or to compute slack.

Because `getShards` does not honor locality options, we have to fall back to `getSelection` on every DB. Doing this on every backup would be ridiculously expensive and wasteful, since we will only end up looking at a newest subset.

What we want is to check availability only for the DBs that are considered for the retention set, and do so only once per DB. Laziness fits like a glove here so I am using `unsafePerformIO`, but it should be easy to do it properly with a State monad if we decide to go down this route.

Differential Revision: D49694989

fbshipit-source-id: 787211dae252fabbbc5b9faeca7b2bb323cbc158