Index the names of ObjC method decls.
Summary:
This diff adds a name field to `ObjcMethodDeclaration` predicate. This is essentially a concatenated version of the `selector` field. However, it does cover some subtle details. For example, the name of an ObjC method with 0 parameters is `foo` whereas the same method with 1 parameter is `foo:`. In multi-arg cases, we capture the full name, `foo:bar:`. This can be constructed using the `selector` field and inspecting `signature` to check the # of parameters and such. For now, Angle is not powerful enough to do this task.

Storing this in the DB doesn't seem so bad since the `ObjcSelector` predicate which is of type `[string]` (no sharing), is < 50MiB.

```
$ glean shell --db fbsource.fbobjc.cxx
Glean Shell, built on 2023-09-27T05:20:29Z, from rev 0aa5e05a0576e41e293689cf8710f2998524ca3b
Using service glean.query.prod
type :help for help.
using database fbsource.fbobjc.cxx/798584286173c90e5d24c825674e8e9828b5b88d
fbsource.fbobjc.cxx> :st cxx1.ObjcSelector
cxx1.ObjcSelector.5
count: 471571
size:  50185460 (47.86 MiB) 0.0785%
```

The name is reported through `Code.Location` in the codemarkup layer, but also used in Symbol ID generation.

This also fixes the codemarkup layer bug where we filter out ObjC method declarations that have more than 1 slot in the selector.

Given an example like:
```
- (void)foo:(int)x bar:(int)y;
^ (2)   ^ (3)
...

[obj foo:1 bar:2]
     ^^^ (1)
```

We now report the jump from (1) successfully through the Codemarkup layer (filtered out before due to the target not having exactly one name).

The result is better since it jumps at all to (2). But ideally we'd want to jump to (3), and same for `bar`. We address that in D49757978.

Reviewed By: donsbot

Differential Revision: D49757801

fbshipit-source-id: 38b9682cca37ac54f52ebb169252f41f46ae6c0e