Index the specific slots of ObjC method selectors.
Summary:
A selector of a ObjC Method has 1 or more slots. For example, a method like `foo` has 1, `foo:` also has 1, `foo: bar:` has 2, etc.

We first index the locations of the slots in the `locations` field of `ObjcMethodDeclaration`. We then refer to these locations with a new `XRefTarget` alternative `ObjcSelectorSlot` which captures the `ObjcMethodDeclaration` and an index into `locations`. This is an important schema design decision in order to be able to support decl <-> def transformation of the methods and be able to index into `locations` of the counterpart.

Reviewed By: simonmar

Differential Revision: D49858407

fbshipit-source-id: a089e80f8a1f7851afd000232ebe519a2f15359c