Add a mechanism to filter available DBs when computing the retention set
Summary:
Currently `listDatabases` lists all the remote DBs available, assuming that they will all eventually become available in an actual query server (after restoring), and that at least some of them are available right now. This approximation has been creating problems recently because:

1. sometimes servers have an outdated remote DB list, or
2. sometimes multiple DBs get published at the same time and none of the DBs returned by `listDatabases` is available right now

We can compute the retention set with full precision if we know which DBs are available elsewhere. That simplifies code a bit as we dont' need to track the remote retention set or to compute slack.
This Diff adds an oracle-like mechanism to find out which DBs are available and avoid situations like 2 above. The oracle is used to ensure that at least `retention_retain_at_least` (available|local) DBs are listed, assuming that there are enough DBs available.

Relevant invariants:
* Only consult the available oracle with DBs that are candidates to be in the retention set.
* To avoid checking all the backup DBs when a tier is bootstrapping and no DBs are available elsewehere yet, we only look at 10X `retention_retain_at_least` DBs. * * Once a DB becomes available we remember it and never ask the oracle again for that DB.

Differential Revision: https://internalfb.com/D49539095

fbshipit-source-id: 73b008bc78b2567d55618034b372696ecb710fc2