list only available shards
Summary:
This provides an external oracle based on `srGetSelection` to filter available DBs.

We want to be careful here: if the oracle doesn't work correctly it could cause a serious service downtime, where only local DBs are advertised. for this reason it is guarded behind a JK "code_indexing/glean:server_filter_available_dbs".

Differential Revision: https://internalfb.com/D49686454

fbshipit-source-id: efda93827d8081ce4b756f16312af38ddb97f38d