list at least retention_retain_at_least available shards (#358)
Summary:
Pull Request resolved: https://github.com/facebookincubator/Glean/pull/358

This provides an external oracle based on `srGetSelection` to filter available DBs.

We want to be careful here: if the oracle doesn't work correctly it could cause a serious service downtime, where only local DBs are advertised. for this reason it is guarded behind a JK "code_indexing/glean:server_filter_available_dbs".

Reviewed By: simonmar

Differential Revision: D49686454

fbshipit-source-id: 5c3f785c8efdd38a58729958019e0f1c0b34578e