Die if the remote DB list is too old (#359)
Summary:
Pull Request resolved: https://github.com/facebookincubator/Glean/pull/359

I investigated using a health check automated restart based on the remote db list age counter, but gave up on the idea because:

* TW does not support HCs driven by counters,
* ServiceHealth CHCs only emit alerts, but don't do automated restarts

Automated restarts are dangerous: do we want Glean to go down if Manifold cannot be reached? Probably better to keep Glean running with an older DB list. To mitigate this:

1. Restart only if Manifold is not believed dead
2. Added a JK to disable the automated restarts altogether

Reviewed By: simonmar

Differential Revision: D49689781

fbshipit-source-id: 961a20343ab16b83d0cf41c4a376b803fd5ef1a6