schema info for a schema id
Summary:
`getSchemaInfo` requires a DB and there are cases where we want to get the schema info for a given schema id, but we dont have an actual DB in hand. While the client can always create a new temporary DB and use it to call `getSchemaInfo`, that's very ugly.

This Diff adds a new Thrift service to request a schema info for a given schema id, as well as backend apis.

Reviewed By: simonmar

Differential Revision: D49913696

fbshipit-source-id: 954b121f53003d91ae1032832196e4d82be344fd