Remove `target@` from units
Summary:
The incremental workflow is using a list of `<filename>` to pass to
`glean create --exclude-file`, but the units created by the indexer
are of the form `<target>@<filename>`, so we're not currently
excluding anything.

This changes the units to be just `<filename>`. Reasons:
1. We don't actually have the targets to hand in the workflow, it
would be a pain to find them
2. When invalidating a file, we really do want to invalidate all
instances of the file even if it is indexed in multiple targets, so
this is the right thing anyway.

Reviewed By: josefs

Differential Revision: D50082251

fbshipit-source-id: 94355a9a435852b73867870ce2889161d9bbfc4f