Fix xrefs for ObjC message expressions.
Summary:
This diff fixes the following scenario. Given an example like:

```
- (void)foo:(int)x bar:(int)y;
^ (2)   ^ (3)      ^ (4)
...

[obj foo:1 bar:2]
           ^^^ (1)
```
We want the jump from (1) to go to (4) rather than (2) as Glean-LSP does today, and rather than (3) as ClangD does today.

We achieve this by capturing the locations of each slot of the selector in the ObjC method declaration, then setting up the `XRefTarget` for us to be able to jump to the specific slot.

Reviewed By: simonmar

Differential Revision: D49757978

fbshipit-source-id: 24d21e4854bf43392188606f2ca4cee3554de8a2