Expand symlinks more selectively
Summary:
We had an issue where
- the path is a symlink inside buck-out
- pointing to a path inside root
- containing a symlink to /mnt/gvfs
- so we didn't resolve the original buck-out path because it resolved
to a location outside root

Instead we want to be more careful about following symlinks: follow
symlinks iteratively and stop if the symlink points outside of root.

Note that this doesn't try to follow symlinks contained within the
path, only when the whole path points to a symlink. I don't know
whether that's a problem or not.

Reviewed By: p0

Differential Revision: D47596061

fbshipit-source-id: b8e9b4c109d478eb849a35cad15f77c1cb42023b