Revert D49682777: Multisect successfully blamed "D49682777: [Codemod][ThriftAnnotationsFBSource] fbcode/glean/" for test or build failures (#372)
Summary:
Pull Request resolved: https://github.com/facebookincubator/Glean/pull/372

This diff is reverting D49682777
D49682777: [Codemod][ThriftAnnotationsFBSource] fbcode/glean/ by generatedunixname226714639793621 has been identified to be causing the following test or build failures:

Tests affected:
- [glean/lang/codemarkup/tests/thrift:tests - cases/xrefs](https://www.internalfb.com/intern/test/844425006128750/)

Here's the Multisect link:
https://www.internalfb.com/multisect/3270522
Here are the tasks that are relevant to this breakage:

We're generating a revert to back out the changes in this diff, please note the backout may land if someone accepts it.

If you believe this diff has been generated in error you may Commandeer and Abandon it.

Reviewed By: simonmar

Differential Revision: D50240775

fbshipit-source-id: a469d03593728873a3b191a8ca7fb17215b3a439