Take the name out of ObjcMethodDeclaration.
Summary:
This diff removes the `name` field out of `ObjcMethodDeclaration` predicate and introduces a `ObjcMethodDeclarationName` predicate instead.

Adding a name field was a [performance problem](https://fb.workplace.com/groups/factsaboutcode/permalink/2098231110526126/) for queries of `ObjcMthodDeclaration` keyed by `Selector`.

Besides that, given only a `Selector`, there's no way to produce the correct corresponding name. This is because `["foo"]` isn't sufficient to know whether it's a `"foo"` or `"foo:"`. This is a mistake in the schema design of `Selector`, but fixing it is not a simple task.

Reviewed By: simonmar

Differential Revision: D50256787

fbshipit-source-id: 0c64e11b13d47643c5e3eac76e087de4ddad9fac