Die if the janitor is stuck for >2h
Summary:
A stuck janitor means new DBs don't get advertised nor restored. After a few hours of this the service is effectively degraded. We have a 20m timeout in place to prevent this (and restart the janitor), but the timeout cannot interrupt FFI calls and we have seen the janitor stuck for >6h in prod.

Let's just kill the process if the janitor is stuck for >2hr. TW will restart the server; there is no other graceful recovery here. There is no need for a HC or alert, since TW will also warn the oncall if this keeps happening.

TODO:
- detect and prevent consecutive 20m timeouts
- collect and log stack traces with Quickstack

Reviewed By: simonmar

Differential Revision: D50362783

fbshipit-source-id: 1e29b59835428cc9aa70fb8dcffac5a4e3e13add