Add a new counter glean.db.<db>.newest
Summary:
In S370809 the janitor was failing and not updating counters, in
particular the `glean.db.<db>.age` counters weren't being updated
which meant that our DB age alerts were not detecting the breakage.

This diff introduces a new counter:

`glean.db.<db>.newest`

The Unix timestamp of the newest available `<db>` on this server.

The idea is that we should base metrics and alerts on this instead of
`.age`, because if the janitor is failing for some reason and not
updating counters, the `.newest` counters will still be correct. We
can turn this into an age using ODS formulas.

Differential Revision: D50448148

fbshipit-source-id: 7e4b56ff550a48bcdf4276aa404eeac456fa3182