change CAPI annotation to specify serialization
Summary:
Currently, all structs (and soon unions) will be on marshal C API by default, unless a field uses `cpp.Type` or `cpp.Adapter`. To opt-in such a struct, someone can currently use `python.MashalCapi`.

However, it would be nice to generalize this to allow forcing serialization if desired. So change the syntax to `python.UseCAPI{ 1: bool serialize = true; }`.

This allows someone to enable an entire module (with `"marshal_python_capi"` target option) but opt-out a particular incompatible struct. Alternatively, can be used to force serialization of a struct that is auto-opted-in by default, e.g., for benchmarking purposes.

Differential Revision: D50390541

fbshipit-source-id: 308e92bd610c26260ae230d7421759c2a3c93775