Better instrument depSlices, which substantially slows Open
Summary:
This step accounts for most of the open time of www.hack.incr.  Poking
at this with rrnewton We've seen latency of 9-10s to completion, which
used to hang silently.  Now it at least says what it's doing when it's
done, giving us a timestamp to base some measurements upon.

It's mapping an FFI DB call `glean_rocksdb_get_unit_id` across a set,
basically.  This looks like it's embarassingly parallel, but I don't
know if there's a way to integrate this backwards into HaXL land via
`asyncOpenDB` or if it would need to be parallelized independently and
locally.  I'm pretty sure it'd be easy to get below 1s though.

Reviewed By: simonmar

Differential Revision: D50856313

fbshipit-source-id: 7dfc861114e66ba0c1e2df56fc53f35615bead9c