Ensure axiom facts are complete when fact dependencies are added.
Summary:
We want to ensure that externally derived facts only depend on facts of predicates that are complete.

For now, we will enforce a weaker restriction, which is that at least axiom facts be complete.
This makes sense because the first predicates to be complete are the ones for the axiom facts.

Reviewed By: pepeiborra

Differential Revision: D48939778

fbshipit-source-id: 690b27ff455bfcfba62709e8dc36d22af213e7d2