Introduce `xrefDecl`.
Summary: This diff introduces a `xrefDecl` helper which handles xref to a decl in a generic way. It cleans up all of the places where we used `clang::dyn_cast` in creating the `XRef` instance. The `XRefVisitor` which is a `clang::ConstDeclVisitor` that encapsulates the xref creation logic to a `clang::Decl`.

Reviewed By: simonmar

Differential Revision: D50919807

fbshipit-source-id: f061ed84c131f7b86d9762f1acf29f4de5c0a05a