Fix logging with strict error handling
Summary:
When a response to a request setting "strict errors" is an error, Glass doesn't log anything. Neither to `glean_glass_server` nor to `glean_glass_server_error_events`. We only noticed this thanks to the user experience metrics for diff-time navigation. The reason is that the logging handler is bypassed by the exception when the stack is unwound. For example, this is what happens for a documentSymbolIndex request:

```
documentSymbolIndex
  -> runRepoFile
    -> withRepoFile
      -> withStrictErrorHandling
        -> exception is thrown here
      -> runLog called here
```

The solution is to move the exception site above the logging site, so that the exception is only ever raised **after logging**.

Reviewed By: phlalx

Differential Revision: D51118704

fbshipit-source-id: 7c4b27c9d7cc4e4aec75346b6f920f5686ee8048