Regression test for incremental C++
Summary:
This is a fully end-to-end test that indexes a base DB, an
incremental DB with some changes, and then checks queries on the
incremental DB against snapshot results. It also tests that most of
the derived predicates are working, including the manual ones.

For this to get the correct results:
* ownership must be working for the base DB
* we have to exclude the correct units
* ownership has to be working for the derived predicates, including
the manual ones (TargetUses and DeclFamilies)

Reviewed By: phlalx

Differential Revision: D51117377

fbshipit-source-id: e67e3cfe60c327702f45d32d44372099c72e04ac