Add query modifiers for searching old/new DBs
Summary:
For writing manual incremental derivers, we will need to explicitly
restrict a predicate search to either the new or old facts. This diff
adds support for that in Angle, e.g.

```
src.File#new _
```

Searches only new facts (facts in the topmost DB in the stack only).

Simlarly, `src.File#old _` would search facts not in the topmost DB.

There's are two new operators in the DSL: `new` and `old`, which do
the obvious thing.

Reviewed By: donsbot

Differential Revision: D51206445

fbshipit-source-id: 294b067cd0045838400999edc88f3e703797156b