Script to compare glass snapshots
Summary:
I have been using this script to figure out whether we get definition references by using a DB indexing the 1-deps of the changeset.

This is currently specialized to compare a Cxx snapshot computed with dependencies against one computed without. But it could be easily extended to other use cases.

Reviewed By: phlalx

Differential Revision: D51314905

fbshipit-source-id: c4f8f63013f86aaa0832b522e685e21d114b496f