Fix "unexpected number of serialized units"
Summary:
This error actually isn't an error. We sometimes index the same source
file multiple times. This is showing up with the fbandroid indexer
currently - probably because it's the only one still using Buck 1
where we run the indexer locally and index multiple targets with the
same indexer process. This probably also started happening when we
removed the target names from the units.

Reviewed By: pepeiborra

Differential Revision: D51346935

fbshipit-source-id: 08b37cb343f34c492db371d723c2a1ced76337a5