Schedule tasks on resume
Summary:
When resuming after a server restart, start any tasks that are now
ready. Normally this will be a no-op, but it makes it easier to
manually hack the meta file because we can put tasks into the
"waiting" state and the server will start them for us. Otherwise we
have to put tasks in the "running" state and wait for a timeout.

Reviewed By: pepeiborra

Differential Revision: D51347258

fbshipit-source-id: 765150f3276600d72362a3af2036998c51680fd1