Store the precomputed slices in the DB
Summary:
The "slice" is a bitmap giving the visibility of each ownership set
in the base DB(s). It is computed from the set of excluded units and
the set contents in the base DB(s).

Computing the slice takes a few seconds: about 6s for www.hack.light,
longer for fbsource.

In this diff I'm serializing the computed slice and storing it in the
DB. Deserializing the slice is way faster than recomputing it each
time, and it only takes a few kB in the DB.

Reviewed By: pepeiborra

Differential Revision: D51308907

fbshipit-source-id: b5a4e119084c4a6efd644c4fe1d0adc9142bf78d