Trace writing to Scuba
Summary:
These traces are useful to better understand the Glean write performance.

Performance impact should be negligible, as the Scuba logging is done in a background thread. If this is a concern I can add a configuration setting to disable this tracing by setting the tracer to be a no-op

Example (created with an earlier version of this): https://fburl.com/scuba/glean_server_events/683r46o4

Reviewed By: simonmar

Differential Revision: D51668959

fbshipit-source-id: 9e2e83cafc56872f9cfeca2a715db081ba962a1f