Fix laziness bugs
Summary:
We were trying to perform ownership substitution in the writer threads
by forcing it with a bang-pattern, but it didn't work because the
field is a `Maybe`.

Fix 1: use `force` to perform substitution when we expect it. This
should improve write performance since we'll do more of the work in
parallel in the writer threads.

Fix 2: move the logging to a place where it will catch errors in
ownership substitution, so we will correctly log any errors emitted by
D51747069.

Reviewed By: phlalx

Differential Revision: D51890059

fbshipit-source-id: cda6116907a7a872252b50c31dee240a8eff70ab