Remove a property test that was false
Summary:
The compose3 property was false, but rapidcheck didn't find it because
the failing case was too unlikely. Making it more likely by reducing
the range of Id made the failure show up.

Why was the test false? because you can only compose substitutions A . B when the domain of B is a subset of the range of A, and the test generator wasn't ensuring that. I don't have the energy to write a correct generator right now.

Reviewed By: donsbot

Differential Revision: D51986356

fbshipit-source-id: 630d3fea14925191e416541452692ad1d1789423