Write facts and ownership in parallel
Summary:
We spend quite a lot of time on the server substituting the ownership
data, so I'm trying to speed things up by doing it in parallel with
committing the facts. This should be safe: writing facts must be
serialized, but there's no reason we can't be substituting and writing
the ownership data at the same time.

Reviewed By: donsbot

Differential Revision: D51978310

fbshipit-source-id: ae5896279e4e4ac28c555f157f4d9c4165b93380