Fall back to snapshots for new files
Summary:
The main use case is glean-lsp when on a file that has been submitted to Phabricator but not yet landed to Prod

Glass already serves snapshots to glean-lsp, the only change required here is to make it more likely that a snapshot will be found for a new file. The problem is that looking up by revision is not a good fit for new files, where the revision is likely to change as the user amends or rebases the stack. Instead, we want to match by filepath and just pick the most recent snapshot available

Reviewed By: phlalx

Differential Revision: D52251986

fbshipit-source-id: d72cee365f47017c02c015c07849493c7ee19664