Add FileEntityXRefRangeSpans to codemarkup
Summary:
Right now the diff sketcher uses FileEntityXRefSpans to get ByteSpans
in favor of FileEntityXRefLocations; the latter yields
incomplete results on partial databases since it can't find a
definition site for many entitites.  This is a particular problem in
Hack, where even the production database is missing definition sites
for some entities.  But many languages (eg C++) use Ranges rather than
ByteSpans, and right now we're dropping those XRefs.  So this diff
introduced FileEntityXRefRangeSpans which does what we actually want.

Note that FileEntityXRefSpans was first introduced by simonmar in
D35141618 to support the sketcher.  So I believe this effectively
deprecates the one predicate in favor of the other.

An alternative is to pick Spans vs Locations by language, but
pepeiborra argues against that and in favor of the present solution
in v12 of D52309715.

Reviewed By: pepeiborra

Differential Revision: D52416817

fbshipit-source-id: 8e84588e234a5a5139d1112713c5c01080238eeb