adding deprecated_default_enum_min_i32: glean
Summary:
The default behaviour for thrift rust enums is to be changed to be spec-conformant.

I'm adding a compatibility flag for the existing thrift rust libraries with enums for safe transition.

This is a part of the first step of the migration described in detail in  https://fb.workplace.com/groups/rust.language/posts/25141981055423839

I did it as a codemod across fbsource, but glean generated TARGETS files need special handling

I generated the changes via:
```
./glean/schema/sync --dev
```

Reviewed By: malanka

Differential Revision: D52368095

fbshipit-source-id: f0d1eb555c6278842a3a6eff27de054dae44dcda