Add a coroutines test.
Summary: This adds a new test for coroutines, and adjusts the coroutines involved in `using-directive4` to test against `FileXRefMap`. This is due to a rather rare situation where `#include <coroutine>` is required to use coroutines. In this case, we can refer to external targets outside of our tests which we don't want to test against.

Reviewed By: simonmar

Differential Revision: D52296129

fbshipit-source-id: 9b6203e995c9f9905bf60bd80c5772c88c34f5b7