Track the full declarations in `declarations.query`.
Summary:
Currently, `declarations.query` tracks the locations of decls, and `declarationlocation.query` tracks the locations and the name of the declaration. That is, `declarationlocation.query` is a superset test at the moment. `declarations.query` used to track the full declarations, but that changed in D39247303.

For D52421611, I want the scope of the declarations to be tracked as well. As such, this diff cleans up the current state a little bit by removing `declarationlocation.query`, and make `declarations.query` track the full declarations again.

Reviewed By: simonmar

Differential Revision: D52421613

fbshipit-source-id: b135ab7a9751639228f0d199fa1b0ce56b6a1953