Fix functions within functions to capture its lexical scope.
Summary:
This diff fixes the scope of a function declared within a function.

We want the following query to be accurate, in order to detect all most-vexing-parse issues in the codebase today:
```
cxx1.FunctionDeclaration { name = { scope = { local = _ } } }
```

This currently doesn't work due to functions declared within functions being considered to be within the global scope.

Reviewed By: simonmar

Differential Revision: D52421611

fbshipit-source-id: d129d6cb3dee29cb572e321a12d334aa1c41b417