Add a test case for implicit ctor initializers.
Summary:
D52296131 was supposed to be a no-behavior change, but it turns out it was a preemptive fix for a bug I just discovered.

D39884391 first introduced capturing xrefs for ctor initializers. This had a problem that we were capturing even the implicit xrefs from defaulted ctors, which was producing bloat in the DB. D44904752 fixed that issue by testing for `isDefaulted()` on the ctor being analyzed. However, this fix only addressed the defaulted ctors, rather than the **implicit xrefs** in general.

An example:

```cpp
struct S {
    S();
//  ^ C

    X x1;
    X x2;
//    ^^ B
};

S::S() : x1() {}
// ^ A
```

In this example, even though `S()` ctor is not defaulted, there still is an implicit `x2()` initialization in the ctor. Without D52296131, we end up with implicit xref from A to B in that example! Expected behavior would be A to C, which is also missing but that's a separate fix. It picks A as the location of the implicit xref since that's not really a better option there. D52296131 fixed this by correctly relying on the Clang AST traversal logic that handles ignoring implicit entities rather than the explicit traversal that was done prior.

Reviewed By: simonmar

Differential Revision: D52581772

fbshipit-source-id: f10d9ef4a26d6cce56b53f7d4e230345ba206ad5